editing layout/ adding logout confirmation

+++ src/app/layout/app.topbar.component.ts
@@ -3,6 +3,8 @@ import { Observable } from 'rxjs';
 import { LayoutService } from 'src/app/layout/service/app.layout.service';
 import { AccessTokenService } from '../modules/auth/shared/services/access-token.service';
 import { AuthService } from '../modules/auth/shared/services/auth.service';
+import { MatDialog } from '@angular/material/dialog';
+import { LogoutConfirmationDialogComponent } from '../logout-confirmation-dialog/logout-confirmation-dialog.component'; // Create a separate component for the confirmation dialog
 
 @Component({
   selector: 'app-topbar',
@@ -15,7 +17,8 @@ export class AppTopbarComponent implements OnInit {
   constructor(
     public layoutService: LayoutService,
     private authService: AuthService,
-    private accessTokenService: AccessTokenService
+    private accessTokenService: AccessTokenService,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -23,7 +26,7 @@ export class AppTopbarComponent implements OnInit {
   }
 
   get isLoggedIn(): boolean {
-    return this.authService.isLoggedInSync(); // Add a synchronous method to get the login status
+    return this.authService.isLoggedInSync();
   }
 
   onMenuButtonClick() {
@@ -39,6 +42,13 @@ export class AppTopbarComponent implements OnInit {
   }
 
   logout() {
-    this.accessTokenService.deleteAccessToken();
+    const dialogRef = this.dialog.open(LogoutConfirmationDialogComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 'confirm') {
+        this.accessTokenService.deleteAccessToken();
+      }
+      // else: Do nothing if 'cancel' or the dialog is closed without making a choice
+    });
   }
 }
